Added Stduent statistic implementation.

+++ TestApp/Services/IStatisticService.cs
@@ -15,7 +15,7 @@ namespace TestApp.Services
         public int GetTeachePreviousWeekLessons();
 
 
-        public int GetStudentAverageLessons(int studentId);
+        public double GetStudentAverageLessons(int studentId);
         public int GetStudentToltalLessons(int studentId);
         public int GetStudentCurrentWeekLessons(int studentId);
         public int GetStudentPreviousWeekLessons(int studentId);
